% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Finding neighbors}
\usage{
find_neighbors(graph)
}
\arguments{
\item{graph}{A symmetric adjacency matrix representing the undirected graph.}
}
\value{
a list. Each list element represents a graph node and contains its neighboring nodes.
}
\description{
Finding neighbors of nodes in a graph.
}
