% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_known.R
\name{test_known}
\alias{test_known}
\title{Local regressions and tests}
\usage{
test_known(graph, data, group, glm.family, Z, progressbar)
}
\arguments{
\item{graph}{A symmetric adjacency matrix representing the undirected graph. Node names must match the variable names in the data.}

\item{data}{A data frame with the observed variables. Variable names in the data must match the node names.}

\item{group}{A vector of group labels for comparison, one entry per observation.}

\item{glm.family}{Description of the distribution and link function to be used in the model.}

\item{progressbar}{Logical, whether to display the progress bar. Default is TRUE.}
}
\value{
A list containing three elements: a matrix of p-values, a matrix of coefficients for the baseline group, and a matrix of coefficients for the differences between the two conditions.
}
\description{
Function to fit node-conditional models for comparison and calculate p-values for the two stages.
}
