% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stages_pval.R
\name{two_stages_pval}
\alias{two_stages_pval}
\title{Two-stage procedure}
\usage{
two_stages_pval(graph, test.pval, method.FDR, method.FWER, alpha)
}
\arguments{
\item{graph}{A symmetric adjacency matrix representing the undirected graph.}

\item{test.pval}{A list returned by the 'test_known' function.}

\item{method.FDR}{Method for adjusting p-values in the first stage to control the FDR.}

\item{method.FWER}{Method for adjusting p-values in the first stage to control the FWER.}

\item{alpha}{The alpha level for conducting the test.}
}
\value{
A matrix of dimension p x (p + 2), where p is the number of nodes, indicating differences in the network.
The columns are as follows:
1. First screening stage: 0 for equality, 1 for nodes with a difference.
2. Difference in the mean parameter of the node in the row: 0 for no difference, 1 for a difference, NA if the first column is 0.
3-(p+2). Differences in the association parameters between the node in the row and nodes in the columns: 0 for no difference, 1 for a difference, NA if not tested.
}
\description{
1. Screening stage
}
\details{
2. Confirmatory stage
}
